관리자쪽 권한검사
페이지 시작시에 추가하려고 했는데 코드가 좀 스파게티 돼서ㅠ 등록, 수정, 삭제 요청시 권한검사 하는걸로 함

+++ src/main/java/OOSE/controller/MemberManagement/DeleteMember.java
@@ -19,12 +19,11 @@ public class DeleteMember extends HttpServlet
     @Override
     protected void doGet(HttpServletRequest req, HttpServletResponse resp) throws IOException
     {
-        //로그인 구현 쪽 병합시 테스트해볼 예정
-//        if (checkAuthority((int) req.getSession().getAttribute("authority")))
-//        {
-//            printAlert("권한이 없습니다",resp);
-//            return;
-//        }
+        if(!checkAuthority((int) req.getSession().getAttribute("authority")))
+        {
+            printAlert("권한이 없습니다",resp);
+            return;
+        }
         Member member = new Member();
         member.setId(req.getParameter("id"));
         if(!dbManager.deleteMemberInfo(member)) //삭제 실패한 경우
